+++ src/pack-docs.js
@@ -50,6 +50,7 @@ function renderPackIndex(packs) {
     <nav class="site-nav">
       <a href="../index.html">Home</a>
       <a href="../quickstart.html">Quickstart</a>
+      <a href="../external-trial.html">External Trial</a>
       <a href="../migration-hub/spring-boot-2-to-3.html">Migration Hub</a>
       <a href="../benchmarks/index.html">Benchmarks</a>
       <a href="../packs/index.html">Packs</a>
@@ -78,6 +79,7 @@ function renderPackPage(pack) {
     <nav class="site-nav">
       <a href="../index.html">Home</a>
       <a href="../quickstart.html">Quickstart</a>
+      <a href="../external-trial.html">External Trial</a>
       <a href="../migration-hub/spring-boot-2-to-3.html">Migration Hub</a>
       <a href="../benchmarks/index.html">Benchmarks</a>
       <a href="index.html">Packs</a>